Convert issue links to use "issue:" URLs
This avoids converted issue tracker data hardcoding links to any
particular version of the issue lists; instead, processes generating
HTML output from the tracker data can most appropriately adjust issue:
URLs to point to an appropriate location for that output.

There is no adjustment implemented for Embedded C issues (intended to
have numbers of the form 0EMB.nn, for nn from 01 to 22) or C Secure
Coding Rules issues (intended to have numbers of the form 0SCR.nn, for
nn from 01 to 02), because there are no links to either that would end
up in issue data (no links at all to specific Embedded C issues, and
the only CSCR links are from the table of contents in that issue
list).